refactor: reorganize usePieChartsData hook and related utilities
