+++ src/hooks/usePieChartsData/const/COLORS.ts
@@ -1,0 +1,13 @@
+export const COLORS = [
+  '#9E0142',
+  '#D53E4F',
+  '#F46D43',
+  '#FDAE61',
+  '#FEE08B',
+  '#FFFFBF',
+  '#E6F598',
+  '#ABDDA4',
+  '#66C2A5',
+  '#3288BD',
+  '#5E4FA2',
+];
