+++ src/hooks/usePieChartsData/types/accumulator.ts
@@ -1,0 +1,3 @@
+import { PieChartDataType } from './pieCharts';
+
+export type Accumulator = Record<string, PieChartDataType>;
